// template-base/frontend/shared/workspace/types.ts
export interface WorkspaceStoreItem extends HierarchicalStoreItem {
  _id: Id<"workspaces">
  name: string
  slug: string
  description?: string
  type: WorkspaceType
  icon?: string
  color: string
  isPublic: boolean
  isMainWorkspace?: boolean
  isLinked?: boolean
  shareDataToParent: boolean
  settings?: {
    allowInvites?: boolean
    allowGuestOnly?: boolean
    requireApproval?: boolean
    defaultRoleId?: Id<"roles">
    icon?: string
  }
  createdBy: Id<"users">
  _creationTime: number
}